requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
python-decouple==3.8
requests==2.31.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0